+++ solidity/contracts/converter/BancorConverterRegistry.sol
@@ -8,8 +8,7 @@ import '../token/interfaces/ISmartTokenController.sol';
 /**
   * @dev The BancorConverterRegistry maintains a list of all active converters in the Bancor Network.
   *
-  * Since converters can be upgraded and thus their address can change, the registry actually
-  * keeps smart tokens internally and not the converters themselves.
+  * Since converters can be upgraded and thus their address can change, the registry actually keeps smart tokens internally and not the converters themselves.
   * The active converter for each smart token can be easily accessed by querying the smart token owner.
   *
   * The registry exposes 3 differnet lists that can be accessed and iterated, based on the use-case of the caller:
@@ -79,6 +78,7 @@ contract BancorConverterRegistry is IBancorConverterRegistry, ContractRegistryCl
     /**
       * @dev adds a converter to the registry
       * anyone can add a converter to the registry, as long as the converter is active and valid
+      * note that a liquidity pool converter can only be added if there's no existing pool with the same reserve configuration
       * 
       * @param _converter converter
     */
@@ -285,6 +285,22 @@ contract BancorConverterRegistry is IBancorConverterRegistry, ContractRegistryCl
     }
 
     /**
+      * @dev returns a list of converters for a given list of smart tokens
+      * this is a utility function that can be used to reduce the number of calls to the contract
+      * 
+      * @param _smartTokens list of smart tokens
+      * @return list of converters
+    */
+    function getConvertersBySmartTokens(address[] _smartTokens) external view returns (address[]) {
+        address[] memory converters = new address[](_smartTokens.length);
+
+        for (uint i = 0; i < _smartTokens.length; i++)
+            converters[i] = ISmartToken(_smartTokens[i]).owner();
+
+        return converters;
+    }
+
+    /**
       * @dev checks whether or not a given converter is valid
       * 
       * @param _converter converter
@@ -293,19 +309,66 @@ contract BancorConverterRegistry is IBancorConverterRegistry, ContractRegistryCl
     function isConverterValid(IBancorConverter _converter) public view returns (bool) {
         ISmartToken token = ISmartTokenController(_converter).token();
 
-        // verifies the the smart token has a supply and that the converter is active
-        if (token.totalSupply() == 0 || token.owner() != address(_converter))
+        // verifies the the smart token has a supply and that the converter is active and enabled
+        if (token.totalSupply() == 0 || token.owner() != address(_converter) || !_converter.conversionsEnabled())
             return false;
 
-        // verifies that the converter holds balance in each of its reserves
         uint reserveTokenCount = _converter.connectorTokenCount();
+        address[] memory reserveTokens = new address[](reserveTokenCount);
+        uint[] memory reserveRatios = new uint[](reserveTokenCount);
+
+        // verifies that the converter holds balance in each of its reserves
         for (uint i = 0; i < reserveTokenCount; i++) {
             IERC20Token reserveToken = _converter.connectorTokens(i);
             if (reserveToken.balanceOf(_converter) == 0)
                 return false;
+            reserveTokens[i] = reserveToken;
+            reserveRatios[i] = connectors(_converter, reserveToken);
         }
 
-        return true;
+        return getLiquidityPoolByReserveConfig(reserveTokens, reserveRatios) == IBancorConverter(0);
+    }
+
+    /**
+      * @dev searches for a liquidity pool with specific reserve tokens / ratios
+      * 
+      * @param _reserveTokens   reserve tokens
+      * @param _reserveRatios   reserve ratios
+      * @return the converter of the liquidity pool, or zero if no such liquidity pool exists
+    */
+    function getLiquidityPoolByReserveConfig(address[] memory _reserveTokens, uint[] memory _reserveRatios) public view returns (IBancorConverter) {
+        // validate input - ensure that the number of reserve tokens match the number of reserve ratio
+        if (_reserveTokens.length == _reserveRatios.length && _reserveTokens.length > 1) {
+
+            // get the smart tokens of the least frequent token (optimization)
+            address[] memory convertibleTokenSmartTokens = getLeastFrequentTokenSmartTokens(_reserveTokens);
+            for (uint i = 0; i < convertibleTokenSmartTokens.length; i++) {
+                ISmartToken smartToken = ISmartToken(convertibleTokenSmartTokens[i]);
+                IBancorConverter converter = IBancorConverter(smartToken.owner());
+
+                // compare reserves
+                uint reserveTokenCount = converter.connectorTokenCount();
+                if (reserveTokenCount == _reserveTokens.length) {
+                    bool identical = true;
+                    for (uint j = 0; j < reserveTokenCount; j++) {
+                        IERC20Token reserveToken = converter.connectorTokens(j);
+                        if (reserveToken != _reserveTokens[j]) {
+                            identical = false;
+                            break;
+                        }
+                        uint256 ratio = connectors(converter, reserveToken);
+                        if (ratio != _reserveRatios[j]) {
+                            identical = false;
+                            break;
+                        }
+                    }
+                    if (identical)
+                        return converter;
+                }
+            }
+        }
+
+        return IBancorConverter(0);
     }
 
     /**
@@ -368,5 +431,44 @@ contract BancorConverterRegistry is IBancorConverterRegistry, ContractRegistryCl
     function removeConvertibleToken(IBancorConverterRegistryData _converterRegistryData, address _convertibleToken, address _smartToken) internal {
         _converterRegistryData.removeConvertibleToken(_convertibleToken, _smartToken);
         emit ConvertibleTokenRemoved(_convertibleToken, _smartToken);
+    }
+
+    function getLeastFrequentTokenSmartTokens(address[] memory _tokens) private view returns (address[] memory) {
+        IBancorConverterRegistryData bancorConverterRegistryData = IBancorConverterRegistryData(addressOf(BANCOR_CONVERTER_REGISTRY_DATA));
+
+        // find the token that has the smallest number of smart tokens
+        uint minSmartTokenCount = bancorConverterRegistryData.getConvertibleTokenSmartTokenCount(_tokens[0]);
+        address[] memory smartTokens = bancorConverterRegistryData.getConvertibleTokenSmartTokens(_tokens[0]);
+        for (uint i = 1; i < _tokens.length; i++) {
+            uint convertibleTokenSmartTokenCount = bancorConverterRegistryData.getConvertibleTokenSmartTokenCount(_tokens[i]);
+            if (minSmartTokenCount > convertibleTokenSmartTokenCount) {
+                minSmartTokenCount = convertibleTokenSmartTokenCount;
+                smartTokens = bancorConverterRegistryData.getConvertibleTokenSmartTokens(_tokens[i]);
+            }
+        }
+        return smartTokens;
+    }
+
+    bytes4 private constant CONNECTORS_FUNC_SELECTOR = bytes4(uint256(keccak256("connectors(address)") >> (256 - 4 * 8)));
+
+    function connectors(address _dest, address _address) private view returns (uint256) {
+        uint256[2] memory ret;
+        bytes memory data = abi.encodeWithSelector(CONNECTORS_FUNC_SELECTOR, _address);
+
+        assembly {
+            let success := staticcall(
+                gas,           // gas remaining
+                _dest,         // destination address
+                add(data, 32), // input buffer (starts after the first 32 bytes in the `data` array)
+                mload(data),   // input length (loaded from the first 32 bytes in the `data` array)
+                ret,           // output buffer
+                64             // output length
+            )
+            if iszero(success) {
+                revert(0, 0)
+            }
+        }
+
+        return ret[1];
     }
 }