+++ solidity/contracts/converter/interfaces/IBancorConverter.sol
@@ -9,6 +9,7 @@ contract IBancorConverter {
     function getReturn(IERC20Token _fromToken, IERC20Token _toToken, uint256 _amount) public view returns (uint256, uint256);
     function convert2(IERC20Token _fromToken, IERC20Token _toToken, uint256 _amount, uint256 _minReturn, address _affiliateAccount, uint256 _affiliateFee) public returns (uint256);
     function quickConvert2(IERC20Token[] _path, uint256 _amount, uint256 _minReturn, address _affiliateAccount, uint256 _affiliateFee) public payable returns (uint256);
+    function conversionsEnabled() public view returns (bool) {this;}
     function conversionWhitelist() public view returns (IWhitelist) {this;}
     function conversionFee() public view returns (uint32) {this;}
     function reserves(address _address) public view returns (uint256, uint32, bool, bool, bool) {_address; this;}