+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### 0.5.8-11 (2019-12-17)
+General
+* Added the BancorNetworkPathFinder contract, now compatible with the new converter registry contract
+
+BancorConverter
+* Added a dedicated getReserveRatio function
+
+BancorConverterRegistry
+* Now enforces only a single liquidity pool for each reserve configuration
+* Disabled converters are now considered invalid (can be removed by anyone)
+* Added a utility function that returns a list of converters for a given list of smart tokens
+
+
 ### 0.5.8-10 (2019-12-12)
 General
 * Minor cleanups